mejora: metodo DELETE envia mensajes de confirmacion. corregir: PUT en los metodos de los controllers MateriaController y ProfesorController ya que no se le seteaba un ID lo cual creaba un nuevo resgitro en lugar de actualizar. agregar: this.id = id en clases domain Materia y Profesor

+++ src/main/java/sv/edu/udb/controller/AlumnoController.java
@@ -30,8 +30,14 @@ public class AlumnoController {
         return repo.save(alumno);
     }
 
+    //----
     @DeleteMapping("/{id}")
-    public void eliminar(@PathVariable Long id){
+    public String eliminar(@PathVariable Long id){
+        if(!repo.existsById(id)){
+            return "Alumno no encontrado";
+        }
+
         repo.deleteById(id);
+        return "Alumno eliminado correctamente";
     }
 }
